Renamed dragon stat block contents files
